rename and put PausableWindow into its own file and add short doc string
