+++ Sources/Preferences/PreferencesWindowController.swift
@@ -6,7 +6,7 @@ public final class PreferencesWindowController: NSWindowController {
 	public init(preferences: [Preferenceable], style: PreferencesStyle = .tabs, crossfadeTransitions: Bool = true) {
 		precondition(!preferences.isEmpty, "You need to set at least one view controller")
 
-		let window = PausableWindow(
+		let window = UserInteractionPausableWindow(
 			contentRect: preferences[0].viewController.view.bounds,
 			styleMask: [
 				.titled,
@@ -40,30 +40,5 @@ public final class PreferencesWindowController: NSWindowController {
 
 	public func hideWindow() {
 		close()
-	}
-}
-
-class PausableWindow: NSWindow {
-	var isUserInteractionEnabled: Bool = true
-
-	let userEventTypes: [NSEvent.EventType] = {
-		var result: [NSEvent.EventType] = [
-			.leftMouseDown, .leftMouseUp, .rightMouseDown, .rightMouseUp, .leftMouseDragged, .rightMouseDragged,
-			.keyDown, .keyUp, .scrollWheel, .tabletPoint, .otherMouseDown, .otherMouseUp, .otherMouseDragged,
-			.gesture, .magnify, .swipe, .rotate, .beginGesture, .endGesture, .smartMagnify, .quickLook, .directTouch]
-
-		if #available(macOS 10.10.3, *) {
-			result.append(.pressure)
-		}
-
-		return result
-	}()
-
-	override func sendEvent(_ event: NSEvent) {
-		if !isUserInteractionEnabled && userEventTypes.contains(event.type) {
-			return
-		}
-
-		super.sendEvent(event)
 	}
 }
